mv all commands and cli logic into one dir
