+++ deal/_cli/__init__.py
@@ -1,0 +1,4 @@
+from ._main import main
+
+
+__all__ = ['main']
