+++ deal/linter/__init__.py
@@ -1,5 +1,6 @@
 # app
 from ._checker import Checker
+from ._stub import StubsManager, generate_stub
 
 
-__all__ = ['Checker']
+__all__ = ['Checker', 'StubsManager', 'generate_stub']
